eth/tracers/internal: move common func into util.go
