+++ eth/tracers/internal/tracetest/statediff_test.go
@@ -32,9 +32,6 @@ import (
 	"github.com/ethereum/go-ethereum/eth/tracers"
 	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/ethereum/go-ethereum/tests"
-
-	// Force-load native and js pacakges, to trigger registration
-	_ "github.com/ethereum/go-ethereum/eth/tracers/native"
 )
 
 // stateDiffTrace is the result of a stateDiffTrace run.
